refactor: LightComponentBase 추가.
