+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/DirectionalLightComponent.cpp
@@ -1,5 +1,5 @@
 #include "DirectionalLightComponent.h"
-#include "Components/SceneComponent.h"
+
 #include "Math/Rotator.h"
 #include "Math/Quat.h"
