+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/DirectionalLightComponent.h
@@ -1,10 +1,9 @@
 #pragma once
 #include "LightComponent.h"
-#include "UObject/ObjectMacros.h"
 
-class UDirectionalLightComponent : public ULightComponentBase
+class UDirectionalLightComponent : public ULightComponent
 {
-    DECLARE_CLASS(UDirectionalLightComponent, ULightComponentBase)
+    DECLARE_CLASS(UDirectionalLightComponent, ULightComponent)
 
 public:
     UDirectionalLightComponent();
