+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/LightComponent.h
@@ -1,22 +1,11 @@
-#pragma once
-#include "Components/SceneComponent.h"
+﻿#pragma once
 
+#include "LightComponentBase.h"
 
-class ULightComponentBase : public USceneComponent
+class ULightComponent : public ULightComponentBase
 {
-    DECLARE_CLASS(ULightComponentBase, USceneComponent)
-
+    
 public:
-    ULightComponentBase();
-    virtual ~ULightComponentBase() override;
-    virtual UObject* Duplicate(UObject* InOuter) override;
-
-    virtual void TickComponent(float DeltaTime) override;
-    virtual int CheckRayIntersection(FVector& rayOrigin, FVector& rayDirection, float& pfNearHitDistance) override;
-protected:
-
-    FBoundingBox AABB;
-
-public:
-    FBoundingBox GetBoundingBox() const {return AABB;}
+    ULightComponent() = default;
+    
 };
