+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/LightComponentBase.cpp
@@ -1,0 +1,34 @@
+#include "LightComponentBase.h"
+#include "UObject/Casts.h"
+
+ULightComponentBase::ULightComponentBase()
+{
+    AABB.max = { 1.f,1.f,0.1f };
+    AABB.min = { -1.f,-1.f,-0.1f };
+}
+
+ULightComponentBase::~ULightComponentBase()
+{
+  
+}
+
+UObject* ULightComponentBase::Duplicate(UObject* InOuter)
+{
+    ThisClass* NewComponent = Cast<ThisClass>(Super::Duplicate(InOuter));
+
+    NewComponent->AABB = AABB;
+
+    return NewComponent;
+}
+
+void ULightComponentBase::TickComponent(float DeltaTime)
+{
+    Super::TickComponent(DeltaTime);
+}
+
+int ULightComponentBase::CheckRayIntersection(FVector& rayOrigin, FVector& rayDirection, float& pfNearHitDistance)
+{
+    bool res = AABB.Intersect(rayOrigin, rayDirection, pfNearHitDistance);
+    return res;
+}
+
