+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/PointLightComponent.h
@@ -1,10 +1,11 @@
 #pragma once
+
 #include "LightComponent.h"
 
-class UPointLightComponent :public ULightComponentBase
+class UPointLightComponent :public ULightComponent
 {
 
-    DECLARE_CLASS(UPointLightComponent, ULightComponentBase)
+    DECLARE_CLASS(UPointLightComponent, ULightComponent)
 public:
     UPointLightComponent();
     virtual ~UPointLightComponent() override;
