+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/Classes/Components/Light/SpotLightComponent.h
@@ -1,10 +1,10 @@
 #pragma once
 #include "LightComponent.h"
 
-class USpotLightComponent :public ULightComponentBase
+class USpotLightComponent :public ULightComponent
 {
 
-    DECLARE_CLASS(USpotLightComponent, ULightComponentBase)
+    DECLARE_CLASS(USpotLightComponent, ULightComponent)
 public:
     USpotLightComponent();
     virtual ~USpotLightComponent();
